Fix: query_start_loc must be int32, not int64 (FA3 scheduler expects Int)
FlashAttention3's get_scheduler_metadata expects cu_seqlens_q as int32.
Normal vLLM path uses self.query_start_loc which is int32.
Changed both prefill and decode query_start_loc from int64 to int32.

+++ vllm/v1/worker/gpu_model_runner.py
@@ -5869,7 +5869,7 @@ class GPUModelRunner(
         # Attention metadata for prefill (single sequence)
         prefill_seq_lens = torch.tensor([prompt_len], dtype=torch.int32, device=device)
         prefill_query_start_loc = torch.tensor(
-            [0, prompt_len], dtype=torch.int64, device=device
+            [0, prompt_len], dtype=torch.int32, device=device
         )
 
         prefill_cm = CommonAttentionMetadata(
@@ -5975,7 +5975,7 @@ class GPUModelRunner(
             )
             query_start_loc = torch.arange(
                 num_active + 1,
-                dtype=torch.int64,
+                dtype=torch.int32,
                 device=device,
             )
